fix: move util runner to backend and rename to public
